poem-agent/requirements.txt
yt-dlp
torch
transformers
openai-whisper
pymupdf
requests
pytesseract
fugashi
spacy
ja_ginza
protobuf
unidic-lite
ipadic